backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-api-python-client==2.140.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-auth==2.35.0
httpx==0.27.2
apscheduler==3.10.4
aiofiles==23.2.1
python-dotenv==1.0.1
pydantic==2.9.2